model/requirements.txt
flask
flask-cors
torch
torchvision
transformers
pillow
requests
cloudinary
gunicorn